Conflicts:
	www/public_ftp/iepe/app/Http/routes.php
	www/public_ftp/iepe/resources/views/layouts/admin-user.blade.php

+++ www/public_ftp/iepe/database/migrations/2014_10_12_000000_create_users_table.php
@@ -36,7 +36,7 @@ class CreateUsersTable extends Migration
         });
 
         Schema::create('aspirantes', function (Blueprint $table) {
-            $table->integer('NOV')->unsigned()->primary();
+            $table->bigInteger('NOV')->unsigned()->primary();
             $table->string('nombre');
             $table->string('apellido');
             $table->string('email')->unique();
@@ -63,7 +63,7 @@ class CreateUsersTable extends Migration
             $table->enum('sector',['privado', 'publico']);
             $table->enum('carrera',['arquitectura', 'diseño']);
             $table->enum('jornada',['matutina', 'vespertina']);
-            $table->integer('NOV')->unsigned();
+            $table->bigInteger('NOV')->unsigned();
             $table->foreign('NOV')->references('NOV')->on('aspirantes');
             $table->timestamps();
         });
